pyproject.toml
[tool.poetry]
name = "dj_blog"
version = "0.1.0"
description = ""
authors = ["surajmt8848 <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
djangorestframework-simplejwt = "^4.6.0"
drf-yasg = "^1.20.0"
pep8 = "^1.7.1"
django-filter = "^2.4.0"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
